+++ app/build.gradle
@@ -37,6 +37,7 @@ dependencies {
 
     implementation 'androidx.recyclerview:recyclerview:1.3.0'
     def activity_version = "1.7.0"
+    def fragment_version = "1.5.6"
     def lifecycle_version = "2.6.1"
 
     implementation 'androidx.core:core-ktx:1.9.0'
@@ -44,6 +45,7 @@ dependencies {
     implementation 'com.google.android.material:material:1.8.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation "androidx.activity:activity-ktx:$activity_version"
+    implementation "androidx.fragment:fragment-ktx:$fragment_version"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_version"
