+++ app/src/main/res/layout/activity_main.xml
@@ -5,101 +5,31 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".activity.MainActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/list"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
+        android:layout_height="match_parent"
         android:clipToPadding="false"
         android:padding="@dimen/common_spaces"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintBottom_toTopOf="@id/tvEditedPost"
-        app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0"
         tools:layout_editor_absoluteX="0dp"
         tools:listitem="@layout/card_post" />
 
-    <androidx.constraintlayout.widget.Barrier
-        android:id="@+id/toolsTop"
-        android:layout_width="wrap_content"
-        app:barrierDirection="top"
-        android:layout_height="wrap_content"
-        app:constraint_referenced_ids="tvEditedPost"/>
-
-    <EditText
-        android:id="@+id/etContent"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:background="@android:color/transparent"
-
-        android:hint="@string/content"
-        android:textColorHint="@color/edit_info"
-        android:importantForAutofill="no"
-        android:inputType="textMultiLine"
-        android:padding="12dp"
-        android:textAlignment="center"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/ivSave"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <ImageButton
-        android:id="@+id/ivSave"
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/add"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:background="@android:color/transparent"
-        android:contentDescription="@string/save"
-        android:padding="8dp"
+        android:layout_margin="@dimen/half_common"
+        android:contentDescription="@string/addpost"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/etContent"
-        app:srcCompat="@drawable/ic_save_48" />
-
-    <ImageButton
-        android:id="@+id/ivCancel"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:background="@android:color/transparent"
-        android:contentDescription="@string/cancel"
-        android:paddingTop="8dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/etContent"
-        app:layout_constraintTop_toTopOf="@id/tvEditedPost"
-        app:srcCompat="@drawable/outline_cancel_24" />
-    <TextView
-        android:id="@+id/tvEditedPost"
-        android:contentDescription="@string/edited"
-        android:text="@string/edited_post"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:padding="8dp"
-        android:textAlignment="center"
-        android:textColor="@color/edit_info"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/ivCancel"
-        app:layout_constraintBottom_toTopOf="@id/tvEditedPostId"
+        app:srcCompat="@drawable/ic_add_24dp" />
 
 
-        />
-    <TextView
-        android:id="@+id/tvEditedPostId"
-        android:contentDescription="@string/editedid"
-        android:text="@string/post_id"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:textAlignment="center"
-        android:padding="8dp"
-        app:layout_constraintEnd_toStartOf="@id/ivSave"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintBottom_toTopOf="@id/etContent"
-        android:textColor="@color/edit_info"
-        />
 
-    <androidx.constraintlayout.widget.Group
-        android:id="@+id/grop_editor"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:constraint_referenced_ids="ivCancel, tvEditedPost, tvEditedPostId"
-        android:visibility="gone"/>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
