+++ app/src/main/res/layout/card_post.xml
@@ -5,7 +5,7 @@
     android:padding="@dimen/common_spaces"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".MainActivity">
+    tools:context=".activity.MainActivity">
 
     <ImageView
         android:id="@+id/ivAvatar"
