+++ app/src/main/res/values/strings.xml
@@ -20,4 +20,6 @@
     <string name="edited">Edited</string>
     <string name="post_author">Post author: </string>
     <string name="posts_id">Post ID: </string>
+    <string name="chooser_share_post">Share post</string>
+    <string name="addpost">addPost</string>
 </resources>
